requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.0
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
python-dotenv==1.0.1